+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 
 ## 0.7.0
 
+- Provide Codelens to `Camel Run with JBang`
+
 ## 0.6.0
 
 - Upgrade Debug Adapter for Apache Camel to 0.6.0